Pluggable storage backend
